+++ src/Controller/AccountController.php
@@ -2,40 +2,88 @@
 
 namespace App\Controller;
 
-use Psr\Log\LoggerInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use App\Entity\User;
+use App\Form\UserRegistrationTypeFormType;
+use App\Security\LoginFormAuthenticator;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\Security\Guard\GuardAuthenticatorHandler;
+use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 
-/**
- * Class AccountController
- * @package App\Controller
- * @IsGranted("ROLE_USER")
- */
-
-class AccountController extends BaseController
+class AccountController extends AbstractController
 {
     /**
-     * @Route("/account", name="account")
+     * @Route("/login", name="login")
      */
-    public function index(LoggerInterface $logger)
+    public function login(AuthenticationUtils $authenticationUtils)
     {
 
-        $logger->debug('Checking account page for '.$this->getUser()->getEmail());
+        // get the login error if there is one
+        $error = $authenticationUtils->getLastAuthenticationError();
+        // last username entered by the user
+        $lastUsername = $authenticationUtils->getLastUsername();
 
-        return $this->render('account/index.html.twig', [
-            'title' => 'AccountController',
+        return $this->render('frontend/account/login.html.twig', [
+            'title' => 'Login',
+            'last_usermail' => $lastUsername,
+            'error' => $error,
         ]);
     }
 
     /**
-     * @Route("/api/account", name="api_account")
+     * @Route("/logout", name="logout")
      */
-    public function accountApi()
+    public function logout()
     {
-        $user = $this->getUser();
+        //TODO: Maybe a redirect?
+    }
 
-        return $this->json($user, 200, [], [
-            'groups' => ['main']
+    /**
+     * @Route("/register", name="register")
+     */
+    public function register(
+        Request $request,
+        UserPasswordEncoderInterface $passwordEncoder,
+        GuardAuthenticatorHandler $guardHandler,
+        LoginFormAuthenticator $formAuthenticator
+        )
+    {
+        $form = $this->createForm(UserRegistrationTypeFormType::class);
+
+        $form->handleRequest($request);
+        if($form->isSubmitted() && $form->isValid()){
+
+            /**
+             * @var User $user
+             */
+            $user = $form->getData();
+
+            $user->setPassword($passwordEncoder->encodePassword(
+                $user,
+                $form['plainPassword']->getData()
+            ));
+
+            if($form['agreeTerms']->getData() === true){
+                $user->agreeTerms();
+            }
+
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($user);
+            $em->flush();
+
+            return $guardHandler->authenticateUserAndHandleSuccess(
+                $user,
+                $request,
+                $formAuthenticator,
+                'main'
+            );
+        }
+
+        return $this->render('frontend/account/register.html.twig', [
+            'title' => 'Registration',
+            'registrationForm' => $form->createView()
         ]);
     }
 }
