+++ src/Controller/BasicPageController.php
@@ -9,27 +9,14 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class BasicPageController extends AbstractController
 {
-
-    /**
-     * @Route("/", name="index")
-     */
-    public function index(ArticleRepository $repository)
-    {
-
-        $articles = $repository->findAllPublishedOrderByNewest();
-
-        return $this->render('basic_page/index.html.twig', [
-            'title' => 'Log',
-            'articles' => $articles
-        ]);
-    }
+    private $imagesBlog = [];
 
     /**
      * @Route("/data", name="data")
      */
     public function data()
     {
-        return $this->render('basic_page/data.html.twig', [
+        return $this->render('frontend/basic_page/data.html.twig', [
             'title' => 'Data'
         ]);
     }
@@ -41,8 +28,28 @@ class BasicPageController extends AbstractController
     {
         $slackClient->sendNotification('patrickeinatz.com','Jemand sieht sich das Archive an!');
 
-        return $this->render('basic_page/archive.html.twig', [
+        return $this->render('frontend/basic_page/archive.html.twig', [
             'title' => 'Archive'
+        ]);
+    }
+
+    /**
+     * @Route("/extras/glueblis", name="extra_glueblis")
+     */
+    public function mons()
+    {
+        $dir = 'images/mons';
+        $fulldir = "{$_SERVER['DOCUMENT_ROOT']}/$dir";
+        $d = @dir($fulldir) or die('Failed opening directory for reading');
+        while(false !== ($entry = @$d->read()))
+        {
+            array_push($this->imagesBlog, "/$dir/$entry");
+        }
+        $d->close();
+
+        return $this->render('frontend/extras/mons.html.twig', [
+            'title' => 'Glüblis',
+            'monsCollection' => $this->imagesBlog,
         ]);
     }
