+++ templates/backend/activity/backend_activity_create.html.twig
@@ -1,0 +1,15 @@
+{% extends 'backend/bare.html.twig' %}
+
+{% block title %}{{ title }}{% endblock %}
+    {% block content_bare%}
+    <h1>Activity - add</h1>
+    {{ form_start(activityForm) }}
+    {{ form_row(activityForm.date) }}
+    {{ form_row(activityForm.training) }}
+    {{ form_row(activityForm.no_alcohol) }}
+    {{ form_row(activityForm.no_candy) }}
+    {{ form_row(activityForm.no_carbs) }}
+    {{ form_row(activityForm.comment) }}
+    <button type="submit" class="btn btn-primary">create</button>
+    {{ form_end(activityForm) }}
+{% endblock %}
