+++ templates/backend/log/backend_log.html.twig
@@ -1,0 +1,22 @@
+{% extends 'backend/bare.html.twig' %}
+
+{% block title %}{{ title }}{% endblock %}
+{% block content_bare%}
+        <h1>Logbook</h1>
+        <div class="row backend-list-element">
+                <div class="col-md-6 backend-list-element-header">Title</div>
+                <div class="col-md-3 backend-list-element-header">Author</div>
+                <div class="col-md-1 backend-list-element-header-center">published</div>
+                <div class="col-md-1 backend-list-element-header-center"></div>
+                <div class="col-md-1 backend-list-element-header-center"></div>
+        </div>
+        {% for article in articles %}
+        <div class="row backend-list-element">
+                <div class="col-md-6 backend-list-element-data">{{ article.title }}</div>
+                <div class="col-md-3 backend-list-element-data">{{ article.author.firstName }}</div>
+                <div class="col-md-1 backend-list-element-data-center"><span class="fa fa-{{ article.isPublished ? 'check-circle' : 'stopwatch' }} text-secondary"></span></div>
+                <div class="col-md-1 backend-list-element-control"><a class="btn btn-sm sub-tool-inline" href="{{ path('backend_log_update', { id: article.id}) }}"><i class="fas fa-pencil-alt"></i></a></div>
+                <div class="col-md-1 backend-list-element-control"><a class="btn btn-sm sub-tool-inline del" href="#"><i class="fas fa-trash-alt"></i></a></div>
+        </div>
+        {% endfor %}
+{% endblock %}
