+++ templates/backend/log/backend_log_create.html.twig
@@ -1,0 +1,9 @@
+{% extends 'backend/bare.html.twig' %}
+
+{% block title %}{{ title }}{% endblock %}
+{% block content_bare%}
+        <h1>Logbook - create</h1>
+        {{ include('backend/log/_form.html.twig', {
+                button_text: 'create'
+        }) }}
+{% endblock %}
