+++ templates/backend/menu/backend_sidebarNavigation.html.twig
@@ -1,0 +1,34 @@
+{% if app.user != null %}
+<div id="MainMenu">
+    <ul class="nav flex-column">
+        <li class="nav-item">
+            <a href="{{ path('backend_account') }}" class="list-group-item list-group-item-action {{ '/account' in app.request.pathInfo  ? 'active'}}"><i class="fas fa-user"></i> Account</a>
+            {% if '/account' in app.request.pathInfo %}
+            <div class="small">
+                <a href="javascript:;" class="list-group-item submenu {{ '/account' in app.request.pathInfo ? 'active'}}"><i class="far fa-edit"></i> Edit</a>
+            </div>
+            {% endif %}
+        </li>
+        <li class="nav-item">
+            <!-- LOGBOOK SECTION -->
+            {% if 'ROLE_ADMIN' in app.user.roles %}
+                <a href="{{ path('backend_log') }}" class="list-group-item list-group-item-action {{ '/log' in app.request.pathInfo  ? 'active'}}{{ '/comment' in app.request.pathInfo  ? 'active'}}"><i class="fas fa-book-dead"></i> Logbook</a>
+                {% if '/log' in app.request.pathInfo or '/comment' in app.request.pathInfo %}
+                    <div class="small">
+                        <a href="{{ path('backend_log_create') }}" class="list-group-item submenu {{ '/log' in app.request.pathInfo  ? 'active'}}{{ '/comment' in app.request.pathInfo  ? 'active'}}"><i class="fa fa-plus-circle"></i> Create</a>
+                        <a href="{{ path('backend_comment') }}" class="list-group-item submenu {{ '/log' in app.request.pathInfo  ? 'active'}}{{ '/comment' in app.request.pathInfo  ? 'active'}}"><i class="far fa-comments"></i> Comments</a>
+                    </div>
+                {% endif %}
+            {% endif %}
+        </li>
+        <li class="nav-item">
+            <a href="{{ path('backend_activity') }}" class="list-group-item list-group-item-action {{ '/activity' in app.request.pathInfo ? 'active'}}"><i class="fas fa-running"></i> Activity</a>
+            {% if '/backend/activity' in app.request.pathInfo %}
+                <div class="small">
+                    <a href="{{ path('backend_activity_create') }}" class="list-group-item submenu {{ '/activity' in app.request.pathInfo ? 'active'}}"><i class="fa fa-plus-circle"></i> Add</a>
+                </div>
+            {% endif %}
+        </li>
+    </ul>
+</div>
+{% endif %}
