--- templates/comment_admin/index.html.twig
@@ -1,49 +1,0 @@
-{% extends 'bare.html.twig' %}
-
-{% block title %}{{ title }}{% endblock %}
-{% block content_bare%}
-        <h1>Comments</h1>
-        <form>
-                <div class="input-group mb-3">
-                        <input type="text"
-                               name="q"
-                               class="form-control"
-                               placeholder="Search..."
-                               value="{{ app.request.query.get('q') }}">
-                        <div class="input-group-append">
-                                <button type="submit"
-                                        class="btn btn-outline-secondary">
-                                        <span class="fa fa-search"></span>
-                                </button>
-                        </div>
-                </div>
-        </form>
-        <span>RESULTS: {{ pagination.getTotalItemCount }}</span>
-        <br><br>
-        <div class="row backend-list-element">
-                <div class="col-md-3 backend-list-element-header">Article</div>
-                <div class="col-md-2 backend-list-element-header">Author</div>
-                <div class="col-md-4 backend-list-element-header">Comment</div>
-                <div class="col-md-1 backend-list-element-header-center">Created At</div>
-                <div class="col-md-1 backend-list-element-header-center">approve</div>
-                <div class="col-md-1 backend-list-element-header-center">delete</div>
-        </div>
-        {% for comment in pagination %}
-        <div class="row backend-list-element">
-                <div class="col-md-3 backend-list-element-data"><a href="{{ path('news', {slug: comment.article.slug}) }}">{{ comment.article.title }}</a></div>
-                <div class="col-md-2 backend-list-element-data">{{ comment.authorName }}</div>
-                <div class="col-md-4 backend-list-element-data">{{ comment.content|truncate(30,true,'...') }}</div>
-                <div class="col-md-1 backend-list-element-data-center">{{ comment.createdAt|ago }}</div>
-                <div class="col-md-1 backend-list-element-control"><a class="btn btn-sm sub-tool-inline" href="#"><i class="fas fa-check-circle"></i></a></div>
-                <div class="col-md-1 backend-list-element-control"><a class="btn btn-sm sub-tool-inline del" href="#"><i class="fas fa-trash-alt"></i></a></div>
-        </div>
-        {% else %}
-        <div class="row backend-list-element">
-                <div class="col-md-12 backend-list-element-data-center">No comments found!</div>
-        </div>
-        {% endfor %}
-        <br><br>
-        <div class="text-center">
-                {{ knp_pagination_render(pagination) }}
-        </div>
-{% endblock %}
