+++ templates/frontend/account/login.html.twig
@@ -1,0 +1,44 @@
+{% extends 'backend/bare.html.twig' %}
+
+{% block title %}{{ title }}{% endblock %}
+
+{% block content_bare%}
+    {% if error %}
+        <div class="alert alert-danger">{{ error.messageKey|trans(error.messageData, 'security') }}</div>
+    {% endif %}
+    <div class="row">
+        <div class="col-md-4"></div>
+        <div class="col-md-4">
+            <h1>Login</h1>
+            <form method="post" class="form" autocomplete="off">
+                <div class="form-group">
+                    <input type="email" class="form-input form-control" id="email" name="email" placeholder="email" value="{{ last_usermail }}" required autofocus/>
+                </div>
+                <div class="form-group">
+                    <input type="password" class="form-input form-control" id="password" name="password" placeholder="password" />
+                </div>
+                {#
+                    If you want to control the URL the user
+                    is redirected to on success (more details below)
+                    <input type="hidden" name="_target_path" value="/account" />
+                #}
+                <div class="form-group">
+                    <small>
+                        <label for="rememberme">
+                            <input type="checkbox" id="rememberme" name="_remember_me">
+                            <span class="form-checkbox-label">Remember me!</span>
+                        </label>
+                    </small>
+                </div>
+                <input type="hidden"
+                       name="_csrf_token"
+                       value="{{ csrf_token('authenticate') }}"
+                >
+                <div class="form-group-centered">
+                    <button type="submit" id="sendlogin" class="btn btn-sm sub-tool">Login</button>
+                </div>
+            </form>
+        </div>
+        <div class="col-md-4"></div>
+    </div>
+{% endblock %}
