+++ templates/frontend/extras/mons.html.twig
@@ -1,0 +1,20 @@
+{% extends 'frontend/base.html.twig' %}
+
+{% block title %}{{ title }}{% endblock %}
+
+{% block jumbotron_md %}<h1>{{ title }}</h1>{% endblock %}
+
+{% block content_md_center %}
+    <div class="row">
+        {% for mon in monsCollection %}
+            {% if mon != "/images/mons/." and mon != "/images/mons/.."%}
+            <div class="col-sm-4" style="text-align: center;">
+                <p>
+                    <img src="{{ mon }}" width="256" class="img-responsive">
+                </p>
+            </div>
+            {% endif %}
+        {% endfor %}
+    </div>
+{% endblock %}
+
