+++ templates/frontend/log/log.html.twig
@@ -1,0 +1,39 @@
+{% extends 'frontend/base.html.twig' %}
+
+{% block title %}{{ title }}{% endblock %}
+{% block stylesheets %}<link rel="stylesheet" href="{{ asset('css/timeline.css') }}">{% endblock %}
+{% block jumbotron_md %}<h1>{{ title }}</h1>{% endblock %}
+
+{% block content_md_center %}
+    <ul class="timeline timeline-centered">
+        {% for article in articles %}
+            <li class="timeline-item">
+                <div class="timeline-info">
+                    <span>{{ article.publishedAt|ago }}</span>
+                </div>
+                <div class="timeline-marker"></div>
+                <div class="timeline-content">
+                    <a href="{{ article.articlePath }}" class="card news_block">
+                        <div class="card-body">
+                            <h4 class="timeline-title">{{ article.title }}</h4>
+                            <p>
+                                {{ article.excerpt}}
+                            </p>
+                            <div class="news_details">
+                                {% if article.tags.count > 0 %}
+                                    {% for tag in article.tags %}
+                                        <span class="badge badge-detail"><i class="fas fa-tags"></i> {{ tag.name }}</span>
+                                    {% endfor %}
+                                    <br>
+                                {% endif %}
+                                <span class="badge badge-detail"><i class="fas fa-user"></i> {{ article.author.firstName }}</span>
+                                <span class="badge badge-detail"><i class="fab fa-hotjar"></i> {{ article.likes }}</span>
+                                <span class="badge badge-detail"><i class="fas fa-comments"></i> {{ article.nonDeletedComments|length }}</span>
+                            </div>
+                        </div>
+                    </a>
+                </div>
+            </li>
+        {% endfor %}
+    </ul>
+{% endblock %}
