+++ templates/frontend/log/log_entry.twig
@@ -1,0 +1,112 @@
+{% extends 'frontend/base.html.twig' %}
+
+{% block title %}{{ article.title }}{% endblock %}
+
+    {% block jumbotron_sm %}<h1>{{ article.title }}</h1>{% endblock %}
+
+{% block content_sm_left %}
+    <div class="news_article_info">
+        <p>
+            <small>{{ article.publishedAt|ago}}</small><br>
+            <small>by {{ article.author.firstName }}</small><br>
+            {% if article.tags.count > 0 %}
+            <br>
+            <small>
+                {% for tag in article.tags %}
+                    <i class="fas fa-tags"></i>&nbsp;{{ tag.name }}<br>
+                {% endfor %}
+            </small>
+            {% endif %}
+        </p>
+        <p>
+            <a href="{{ path('log_toggleLits', {slug: article.slug}) }}" class="fab fa-hotjar cold like-article js-like-article"></a> <span class="js-like-article-count">{{ article.likes }}</span>
+        </p>
+    </div>
+{% endblock %}
+
+{% block content_sm_center %}
+    <div class="row">
+        <div class="col-md-2">
+            <i class="fas fa-play-circle fa-4x"
+               id="playbutton"
+               style="float:left; margin-top: 5px; color: #cccccc;"
+               onclick="togglePlay()"
+               onmouseover="mouseOver()"
+               onmouseout="mouseOut()"
+            ></i></div>
+        <div class="col-md-10"><div id="waveform"></div></div>
+    </div>
+
+    <p>
+    {{ article.content|cached_markdown }}
+    </p>
+    <h5>{{ article.nonDeletedComments|length }} Comments</h5>
+    <hr>
+    <div class="row">
+        {% for comment in article.nonDeletedComments %}
+        <p>
+            <div class="col-md-1">
+                IMG
+            </div>
+            <div class="col-md-11">
+                <span class="bolder">{{ comment.authorName }}</span> ~ <span><small> {{ comment.createdAt|ago }} </small></span><br>
+                <span>{{ comment.content }}</span>
+            </div>
+        </p>
+        {% endfor %}
+    </div>
+{% endblock %}
+{% block javascripts %}
+<script src="{{ asset('js/wavesurfer.js') }}"></script>
+<script src="{{ asset('js/wavesurfer.cursor.js') }}"></script>
+<script type="application/javascript">
+    var wavesurfer = WaveSurfer.create({
+        container: '#waveform',
+        waveColor: '#cccccc',
+        progressColor: '#1d1e1f',
+        height: 75,
+        cursorWidth: 1,
+        cursorColor: '#cccccc',
+        hideScrollbar: true,
+        responsive: true,
+        scrollParent: true,
+        plugins: [
+            WaveSurfer.cursor.create({
+                showTime: true,
+                opacity: 1,
+                hidden: true,
+                customShowTimeStyle: {
+                    'background-color': '#1d1e1f',
+                    color: '#fff',
+                    padding: '2px',
+                    'font-size': '12px'
+                }
+            })
+        ]
+    });
+
+    wavesurfer.load('{{ asset('/uploads/audio_files/' ~  article.audioFile ) }}');
+
+    function togglePlay()
+    {
+        if(wavesurfer.isPlaying()){
+            wavesurfer.pause();
+            document.getElementById("playbutton").className = "fas fa-play-circle fa-4x";
+        }
+        else {
+            wavesurfer.play();
+            document.getElementById("playbutton").className = "fas fa-pause-circle fa-4x";
+        }
+
+    }
+
+    function mouseOver() {
+        document.getElementById("playbutton").style.color = "#1d1e1f";
+    }
+
+    function mouseOut() {
+        document.getElementById("playbutton").style.color = "#cccccc";
+    }
+
+</script>
+{% endblock %}
