--- templates/login/register.html.twig
@@ -1,68 +1,0 @@
-{% extends 'bare.html.twig' %}
-{% block title %}{{ title }}{% endblock %}
-{% form_theme registrationForm _self %}
-
-{% block form_row %}
-    {%- set widget_attr = {} -%}
-    {%- if help is not empty -%}
-        {%- set widget_attr = {attr: {'aria-describedby': id ~"_help"}} -%}
-    {%- endif -%}
-
-    {{- form_label(form, null, {
-        label_attr: {class: 'sr-only'}
-    }) -}}
-    {{- form_errors(form) -}}
-    {{- form_widget(form, widget_attr) -}}
-    {{- form_help(form) -}}
-
-{% endblock %}
-
-{% block _user_registration_type_form_agreeTerms_row %}
-    <div class="form-group">
-        {{ form_errors(form) }}
-        <small>
-            <label for="agreeTerms">
-                <input type="checkbox" id="agreeTerms" name="{{ full_name }}">
-                <span class="form-checkbox-label">Agree the terms I for sure read</span>
-            </label>
-        </small>
-    </div>
-{% endblock %}
-
-{% block content_bare%}
-<div class="row">
-    <div class="col-md-4"></div>
-    <div class="col-md-4">
-        <h1>Registration</h1>
-        {{ form_start(registrationForm, {attr: {class: 'form', autocomplete: 'off'} }) }}
-            {{ form_row(registrationForm.firstName,
-            {attr: {
-                class: 'form-input form-control',
-                placeholder: 'username'
-            }
-            }) }}
-
-
-            {{ form_row(registrationForm.email,
-                {attr: {
-                    class: 'form-input form-control',
-                    placeholder: 'email'
-                }
-            }) }}
-
-            {{ form_row(registrationForm.plainPassword,
-                {attr: {
-                    class: 'form-input form-control',
-                    placeholder: 'password'
-                }
-            }) }}
-
-            {{ form_row(registrationForm.agreeTerms) }}
-        <button type="submit" class="btn btn-sm sub-tool">
-            Register
-        </button>
-        {{ form_end(registrationForm) }}
-    </div>
-    <div class="col-md-4"></div>
-</div>
-{% endblock %}
