--- templates/news_admin/edit.html.twig
@@ -1,9 +1,0 @@
-{% extends 'bare.html.twig' %}
-
-{% block title %}{{ title }}{% endblock %}
-{% block content_bare%}
-        <h1>Logbook - edit</h1>
-        {{ include('news_admin/_form.html.twig', {
-                button_text: 'update'
-        }) }}
-{% endblock %}
