--- templates/news_admin/new.html.twig
@@ -1,9 +1,0 @@
-{% extends 'bare.html.twig' %}
-
-{% block title %}{{ title }}{% endblock %}
-{% block content_bare%}
-        <h1>Logbook - create</h1>
-        {{ include('news_admin/_form.html.twig', {
-                button_text: 'create'
-        }) }}
-{% endblock %}
